+++ oap-server/server-receiver-plugin/otel-receiver-plugin/src/test/java/org/apache/skywalking/oap/server/receiver/otel/otlp/OpenTelemetryMetricRequestProcessorTest.java
@@ -48,7 +48,7 @@ public class OpenTelemetryMetricRequestProcessorTest {
 
     @BeforeEach
     public void setUp() {
-        manager = new ModuleManager();
+        manager = new ModuleManager("Test");
         config = new OtelMetricReceiverConfig();
         metricRequestProcessor = new OpenTelemetryMetricRequestProcessor(manager, config);
         nodeLabels = new HashMap<>();